Fix compiler warnings and some vulnerabilities
* add Q_strlcpy
* add MSG_ReadString_Safe which writes directly to user-provided buffer
  of given size
* add Sys_DPrintf for printing only in developer mode
* add some macros for compiler features (Q_UNLIKELY, Q_FALLTHROUGH,
  Q_NORETURN, etc.)
* add Sys_Printf_f, Sys_DPrintf_f and Sys_Error_f macros for wrapping
  text with function name
* add Q_vsnprintf and Q_snprintf resolving to proper function call
  depending on OS
* add Con_QuakeBar for generating console bar of given width
* increase MAX_OSPATH to 4096
* fix main() to respect -nostdout flag
* vid_x.c: replace printf with Sys_Printf/Sys_DPrintf
* CMakeLists.txt: add option to enable static analyzer on GCC

+++ src/d_modech.c
@@ -81,9 +81,9 @@ void D_ViewChanged (void)
 			r_refdef.vrectbottom - (d_pix_max << d_y_aspect_shift);
 
 	{
-		int		i;
+		size_t	i;
 
-		for (i=0 ; i<vid.height; i++)
+		for (i = 0; i < vid.height; i++)
 		{
 			d_scantable[i] = i*rowbytes;
 			zspantable[i] = d_pzbuffer + i*d_zwidth;
@@ -92,4 +92,6 @@ void D_ViewChanged (void)
 
 	D_Patch ();
 }
+
+// vim: set noexpandtab tabstop=4 shiftwidth=4 :
